+++ engine/hid/src/test/wscript
@@ -21,11 +21,13 @@ def build(bld):
                       target   = 'test_hid',
                       source   = 'test_hid.cpp')
 
-    platform_lib = ['PLATFORM']
-    graphics_lib = ['GRAPHICS', 'DMGLFW']
+    platform_lib = []
+    graphics_lib = []
     graphics_lib_symbols = []
 
     if platform_supports_feature(bld.env.PLATFORM, 'opengl', {}):
+        platform_lib = ['PLATFORM']
+        graphics_lib = ['GRAPHICS', 'DMGLFW']
         graphics_lib += ['OPENGL']
         graphics_lib_symbols += ['GraphicsAdapterOpenGL']
 
@@ -37,6 +39,7 @@ def build(bld):
     if len(graphics_lib_symbols) == 0:
         graphics_lib = ['GRAPHICS_NULL']
         graphics_lib_symbols = ['GraphicsAdapterNull']
+        platform_lib = ['PLATFORM_NULL']
 
     if bld.env.PLATFORM in ('arm64-nx64') and 'DMGLFW' in graphics_lib:
         graphics_lib.remove('DMGLFW')
