+++ engine/platform/src/platform_window_null.cpp
@@ -80,10 +80,13 @@ namespace dmPlatform
 
     uint32_t GetWindowStateParam(HWindow window, WindowState state)
     {
-        if (state == WINDOW_STATE_OPENED)
+        switch(state)
         {
-            return window->m_WindowOpened;
+            case WINDOW_STATE_OPENED: return window->m_WindowOpened;
+            case WINDOW_STATE_FSAA_SAMPLES: return window->m_CreateParams.m_Samples;
+            default:break;
         }
+
         return 0;
     }
