+++ engine/script/src/wscript
@@ -85,4 +85,5 @@ def build(bld):
     if 'web' in bld.env.PLATFORM:
         bld.install_files('${PREFIX}/lib/%s/js' % bld.env.PLATFORM, '../lib/js/library_script.js', postpone = False)
 
-    bld.recurse('test')
+    if not waflib.Options.options.skip_build_tests:
+        bld.recurse('test')
