+++ external/bullet3d/README.md
@@ -1,0 +1,14 @@
+#
+
+Build steps:
+
+* configure
+* build
+* package
+
+Example:
+
+    $ PREFIX=$DYNAMO_HOME waf configure --platform=x86_64-win32
+    $ waf install
+
+It will install to `<defold>\packages\bullet-2.77-x86_64-win32.tar.gz`
